fix dfg-viewer permission system
